Develop

+++ src/main/java/com/ebremer/beakgraph/core/GSPODictionary.java
@@ -10,16 +10,16 @@ import org.apache.jena.graph.Node;
 public interface GSPODictionary {
     
     public long locateGraph(Node element);
-    public Object extractGraph(int id);
+    public Object extractGraph(long id);
     
     public long locateSubject(Node element);
-    public Object extractSubject(int id);
+    public Object extractSubject(long id);
     
     public long locatePredicate(Node element);
-    public Object extractPredicate(int id);
+    public Object extractPredicate(long id);
     
     public long locateObject(Node element);
-    public Object extractObject(int id);
+    public Object extractObject(long id);
     
     public Stream<Node> streamGraphs();
     public Stream<Node> streamSubjects();
